refactor(web): restructure app-sidebar with component decomposition and folder organization
- Decompose app-info.tsx (66/100 complexity, 475 lines) into 6 focused modules:
  extracted use-app-info-actions hook, app-mode-labels utility, and 3 sub-components
  (trigger, detail-panel, modals), reducing main component to 5/100 complexity, 76 lines
- Reorganize flat directory into functional folders: app-info/, nav-link/, dataset-info/__tests__/
- Consolidate scattered test files into __tests__ directories
- Rename navLink to nav-link following kebab-case convention
- Remove unused legacy assets (style.module.css, completion.png, expert.png)
- Replace 5-level nested ternary for app mode labels with shared switch utility
- Unify 6 independent modal boolean states into single activeModal discriminated type
