+++ web/app/(commonLayout)/app/(appDetailLayout)/[appId]/layout-main.tsx
@@ -1,6 +1,6 @@
 'use client'
 import type { FC } from 'react'
-import type { NavIcon } from '@/app/components/app-sidebar/navLink'
+import type { NavIcon } from '@/app/components/app-sidebar/nav-link'
 import type { App } from '@/types/app'
 import {
   RiDashboard2Fill,
