+++ web/app/components/app-sidebar/__tests__/sidebar-animation-issues.spec.tsx
