+++ web/app/components/app-sidebar/__tests__/text-squeeze-fix-verification.spec.tsx
