+++ web/app/components/app-sidebar/dataset-sidebar-dropdown.tsx
@@ -1,4 +1,4 @@
-import type { NavIcon } from './navLink'
+import type { NavIcon } from './nav-link'
 import type { DataSet } from '@/models/datasets'
 import {
   RiMenuLine,
@@ -21,7 +21,7 @@ import Divider from '../base/divider'
 import Effect from '../base/effect'
 import ExtraInfo from '../datasets/extra-info'
 import Dropdown from './dataset-info/dropdown'
-import NavLink from './navLink'
+import NavLink from './nav-link'
 
 type DatasetSidebarDropdownProps = {
   navigation: Array<{
@@ -107,12 +107,12 @@ const DatasetSidebarDropdown = ({
                 </div>
                 <div className="flex flex-col gap-y-1 pb-0.5">
                   <div
-                    className="system-md-semibold truncate text-text-secondary"
+                    className="truncate text-text-secondary system-md-semibold"
                     title={dataset.name}
                   >
                     {dataset.name}
                   </div>
-                  <div className="system-2xs-medium-uppercase text-text-tertiary">
+                  <div className="text-text-tertiary system-2xs-medium-uppercase">
                     {isExternalProvider && t('externalTag', { ns: 'dataset' })}
                     {!!(!isExternalProvider && dataset.doc_form && dataset.indexing_technique) && (
                       <div className="flex items-center gap-x-2">
@@ -123,7 +123,7 @@ const DatasetSidebarDropdown = ({
                   </div>
                 </div>
                 {!!dataset.description && (
-                  <p className="system-xs-regular line-clamp-3 text-text-tertiary first-letter:capitalize">
+                  <p className="line-clamp-3 text-text-tertiary system-xs-regular first-letter:capitalize">
                     {dataset.description}
                   </p>
                 )}
