--- web/app/components/app-sidebar/style.module.css
@@ -1,11 +1,0 @@
-.sidebar {
-    border-right: 1px solid #F3F4F6;
-}
-
-.completionPic {
-background-image: url('./completion.png')
-}
-
-.expertPic {
-background-image: url('./expert.png')
-}
